make module Factory platform specific
