+++ cmd/system-probe/api/module/common.go
@@ -10,21 +10,11 @@ import (
 	"errors"
 
 	"google.golang.org/grpc"
-
-	"github.com/DataDog/datadog-agent/cmd/system-probe/config"
 )
 
 // ErrNotEnabled is a special error type that should be returned by a Factory
 // when the associated Module is not enabled.
 var ErrNotEnabled = errors.New("module is not enabled")
-
-// Factory encapsulates the initialization of a Module
-type Factory struct {
-	Name             config.ModuleName
-	ConfigNamespaces []string
-	Fn               func(cfg *config.Config) (Module, error)
-	NeedsEBPF        func() bool
-}
 
 // Module defines the common API implemented by every System Probe Module
 type Module interface {
