+++ cmd/system-probe/api/module/loader_linux.go
@@ -12,15 +12,24 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/ebpf"
 )
 
-func preRegister(_ *config.Config, isEBPFRequired bool) error {
-	if isEBPFRequired {
+func isEBPFRequired(factories []Factory) bool {
+	for _, f := range factories {
+		if f.NeedsEBPF() {
+			return true
+		}
+	}
+	return false
+}
+
+func preRegister(_ *config.Config, moduleFactories []Factory) error {
+	if isEBPFRequired(moduleFactories) {
 		return ebpf.Setup(ebpf.NewConfig())
 	}
 	return nil
 }
 
-func postRegister(_ *config.Config, isEBPFRequired bool) error {
-	if isEBPFRequired {
+func postRegister(_ *config.Config, moduleFactories []Factory) error {
+	if isEBPFRequired(moduleFactories) {
 		ebpf.FlushBTF()
 	}
 	return nil
