+++ cmd/system-probe/modules/crashdetect_windows.go
@@ -11,12 +11,13 @@ import (
 	"fmt"
 	"net/http"
 
+	"google.golang.org/grpc"
+
 	"github.com/DataDog/datadog-agent/cmd/system-probe/api/module"
 	"github.com/DataDog/datadog-agent/cmd/system-probe/config"
 	"github.com/DataDog/datadog-agent/cmd/system-probe/utils"
 	"github.com/DataDog/datadog-agent/pkg/collector/corechecks/system/wincrashdetect/probe"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
-	"google.golang.org/grpc"
 )
 
 // WinCrashProbe Factory
@@ -33,9 +34,6 @@ var WinCrashProbe = module.Factory{
 			WinCrashProbe: cp,
 		}, nil
 	},
-	NeedsEBPF: func() bool {
-		return false
-	}
 }
 
 var _ module.Module = &winCrashDetectModule{}
