+++ cmd/system-probe/modules/eventmonitor.go
@@ -8,11 +8,8 @@
 package modules
 
 import (
-	"runtime"
-
 	"github.com/DataDog/datadog-agent/cmd/system-probe/api/module"
 	"github.com/DataDog/datadog-agent/cmd/system-probe/config"
-	coreconfig "github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/DataDog/datadog-agent/pkg/eventmonitor"
 	emconfig "github.com/DataDog/datadog-agent/pkg/eventmonitor/config"
 	"github.com/DataDog/datadog-agent/pkg/network/events"
@@ -22,65 +19,59 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
 
-// EventMonitor - Event monitor Factory
-var EventMonitor = module.Factory{
-	Name:             config.EventMonitorModule,
-	ConfigNamespaces: []string{"event_monitoring_config", "runtime_security_config"},
-	Fn: func(sysProbeConfig *config.Config) (module.Module, error) {
-		emconfig := emconfig.NewConfig(sysProbeConfig)
+var eventMonitorModuleConfigNamespaces = []string{"event_monitoring_config", "runtime_security_config"}
 
-		secconfig, err := secconfig.NewConfig()
+func createEventMonitorModule(sysProbeConfig *config.Config) (module.Module, error) {
+	emconfig := emconfig.NewConfig(sysProbeConfig)
+
+	secconfig, err := secconfig.NewConfig()
+	if err != nil {
+		log.Errorf("invalid probe configuration: %v", err)
+		return nil, module.ErrNotEnabled
+	}
+
+	opts := eventmonitor.Opts{}
+	secmoduleOpts := secmodule.Opts{}
+
+	// adapt options
+	if secconfig.RuntimeSecurity.IsRuntimeEnabled() {
+		secmodule.UpdateEventMonitorOpts(&opts, secconfig)
+	} else {
+		secmodule.DisableRuntimeSecurity(secconfig)
+	}
+
+	evm, err := eventmonitor.NewEventMonitor(emconfig, secconfig, opts)
+	if err != nil {
+		log.Errorf("error initializing event monitoring module: %v", err)
+		return nil, module.ErrNotEnabled
+	}
+
+	if secconfig.RuntimeSecurity.IsRuntimeEnabled() {
+		cws, err := secmodule.NewCWSConsumer(evm, secconfig.RuntimeSecurity, secmoduleOpts)
 		if err != nil {
-			log.Errorf("invalid probe configuration: %v", err)
-			return nil, module.ErrNotEnabled
+			return nil, err
 		}
+		evm.RegisterEventConsumer(cws)
+		log.Info("event monitoring cws consumer initialized")
+	}
 
-		opts := eventmonitor.Opts{}
-		secmoduleOpts := secmodule.Opts{}
-
-		// adapt options
-		if secconfig.RuntimeSecurity.IsRuntimeEnabled() {
-			secmodule.UpdateEventMonitorOpts(&opts, secconfig)
-		} else {
-			secmodule.DisableRuntimeSecurity(secconfig)
-		}
-
-		evm, err := eventmonitor.NewEventMonitor(emconfig, secconfig, opts)
+	if emconfig.NetworkConsumerEnabled {
+		network, err := events.NewNetworkConsumer(evm)
 		if err != nil {
-			log.Errorf("error initializing event monitoring module: %v", err)
-			return nil, module.ErrNotEnabled
+			return nil, err
 		}
+		evm.RegisterEventConsumer(network)
+		log.Info("event monitoring network consumer initialized")
+	}
 
-		if secconfig.RuntimeSecurity.IsRuntimeEnabled() {
-			cws, err := secmodule.NewCWSConsumer(evm, secconfig.RuntimeSecurity, secmoduleOpts)
-			if err != nil {
-				return nil, err
-			}
-			evm.RegisterEventConsumer(cws)
-			log.Info("event monitoring cws consumer initialized")
+	if emconfig.ProcessConsumerEnabled {
+		process, err := procconsumer.NewProcessConsumer(evm)
+		if err != nil {
+			return nil, err
 		}
+		evm.RegisterEventConsumer(process)
+		log.Info("event monitoring process-agent consumer initialized")
+	}
 
-		if emconfig.NetworkConsumerEnabled {
-			network, err := events.NewNetworkConsumer(evm)
-			if err != nil {
-				return nil, err
-			}
-			evm.RegisterEventConsumer(network)
-			log.Info("event monitoring network consumer initialized")
-		}
-
-		if emconfig.ProcessConsumerEnabled {
-			process, err := procconsumer.NewProcessConsumer(evm)
-			if err != nil {
-				return nil, err
-			}
-			evm.RegisterEventConsumer(process)
-			log.Info("event monitoring process-agent consumer initialized")
-		}
-
-		return evm, err
-	},
-	NeedsEBPF: func() bool {
-		return runtime.GOOS != "windows" && !coreconfig.SystemProbe.GetBool("runtime_security_config.ebpfless.enabled")
-	},
+	return evm, err
 }
