+++ cmd/system-probe/modules/network_tracer_windows.go
@@ -1,0 +1,20 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build windows
+
+package modules
+
+import (
+	"github.com/DataDog/datadog-agent/cmd/system-probe/api/module"
+	"github.com/DataDog/datadog-agent/cmd/system-probe/config"
+)
+
+// NetworkTracer is a factory for NPM's tracer
+var NetworkTracer = module.Factory{
+	Name:             config.NetworkTracerModule,
+	ConfigNamespaces: networkTracerModuleConfigNamespaces,
+	Fn:               createNetworkTracerModule,
+}
